feat: add functions and member variables to nine lives so that its effect can be modded
This is accomplished by adding a moddable onProc function and some member variables to `perk_nine_lives.nut` that modders can edit to change the functionality. The PerkOrder of the nine lives perk is set to VeryLast + 10000 so that the DamageOverTime SkillType is removed from those skills at the very end so that any other mod or skill trying to access DamageOverType skills during onDamageReceived get the proper skill list.