Refactor: revert armor pen changes made to two-handed wooden flail
The commit 80acb0ac10655d4ef5a3786f1cf5af56aa19cce1 refactored the ArmorPen of the two-handed wooden flail without changing the actual value.
There was no explanation for adding the additional line and it is only confusing at this point.
This commit refactors that change back to how the vanilla weapon is coded